Add `php.ini` configuration manipulation.
Using `php_ini` variable, which can be declared anywhere - as a list -
it is possible to manipulate any desired PHP configuration. The PHP
configuration file is defined in the OS configuration file.

Relevant default configuration, specially for development purposes, are
defined in an internal variable which is "merged" with the
user-variable, ensuring a stable configuration environment. The default
variable can be overwritten also, in any place, as any other default
variable.