Move php5 role to root directory.
Providing the structure of a ansible-galaxy package instead of a
repository of many roles. Easing maintenance and allowing third party
developers to also use that role on their development/deploy
environment.

No meta information for ansible-galaxy publication is added, not the
README was updated with usefull information about the package.