Fixes FPM configuration.
Instead of using Ansibles `ini_file` command, which does not allow us to
define the order which configuration options are set on FPM's INI
configuration file, we now use a template with all the default options
in it.

All FPM configuration was moved under `hwr_fpm_config` variable, which
is a list of allowed directives and none of them is required since the
defaults are set on template.

Using the template, makes unecessary the task of compiling the options
for FPM configuration.

Fixes #7 (on GitHub).