Support PHP 5.4 on Debian squeeze.
Separates debian related tasks on a different role file (tasks/Debian.yml),
creates tasks for importing APT keys and repositories to be used for
installation.

A different varaible file (Debian-squeeze) is used after importing the
`vars/Debian.yml` file adding configurations for different repositories
to be used.

On Vagrantfile, a new machine was created using Debian Wheezy so we can
proper test the playbook on a different environment.