Support php_versions configuration.
Enabling specifying a different version of PHP form the one that
defaults to the distribution. Not all versions are supported by every
distribution, a version compatibility matrix is available on the README.

Each version and lsb distribution has its own configuration file, if a
configuration file is not present the provisioning will fail (because
the file doesn't exist).

This adds a simple shell script that combined with the changes on
Vagrantfile (a machine for each supported PHP version and distribution)
can be considered as our test suite, since our CI does not have multiple
OSes and the ones I tested were crap.