Enable configuration of multiple php.ini.
When using an HTTP server, different PHP files may be present and we
also want to apply out changes to those files as well.

This commit adds a `php_ini_files` variable which lists files that
symlink to the CLI `php.ini` file. PHP FPM configuration is already
managed bu the role.